removing unneeded class
