+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/search/builder/SearchQueryExecutors.java
@@ -20,6 +20,7 @@
 package ca.uhn.fhir.jpa.search.builder;
 
 import ca.uhn.fhir.jpa.model.dao.JpaPid;
+import ca.uhn.fhir.jpa.search.builder.models.ResolvedSearchQueryExecutor;
 import org.apache.commons.lang3.Validate;
 
 import java.util.Iterator;
@@ -57,40 +58,6 @@ public class SearchQueryExecutors {
 		return new ResolvedSearchQueryExecutor(rawPids);
 	}
 
-	/**
-	 * Adapt bare Iterator to our internal query interface.
-	 */
-	static class ResolvedSearchQueryExecutor implements ISearchQueryExecutor {
-		private final Iterator<Long> myIterator;
-
-		ResolvedSearchQueryExecutor(Iterable<Long> theIterable) {
-			this(theIterable.iterator());
-		}
-
-		ResolvedSearchQueryExecutor(Iterator<Long> theIterator) {
-			myIterator = theIterator;
-		}
-
-		@Nonnull
-		public static ResolvedSearchQueryExecutor from(List<Long> rawPids) {
-			return new ResolvedSearchQueryExecutor(rawPids);
-		}
-
-		@Override
-		public boolean hasNext() {
-			return myIterator.hasNext();
-		}
-
-		@Override
-		public Long next() {
-			return myIterator.next();
-		}
-
-		@Override
-		public void close() {
-			// empty
-		}
-	}
 
 	public static ISearchQueryExecutor from(Iterator<JpaPid> theIterator) {
 		return new JpaPidQueryAdaptor(theIterator);
